debugging

+++ src/testing/FileWriter.cu
@@ -13,7 +13,7 @@ void WriteResultsToFile(std::string file_name, int fft_length, __half* data){
     for(int i=0; i<fft_length; i++){
       float re = data[i];
       float im = data[i + fft_length];
-      float x = i/fft_length;
+      float x = static_cast<double>(i)/static_cast<double>(fft_length);
       myfile << x << " " << re << " " << im << "\n";
     }
     myfile.close();
